Moves Float Fetcher to Services
