+++ src/lib/services/float_fetcher.ts
